cleaning the previous thunk code and fully utilising rtk query , structured the dashboard
