+++ package.json
@@ -13,12 +13,11 @@
     "@headlessui/react": "^2.0.3",
     "@heroicons/react": "^2.1.3",
     "@reduxjs/toolkit": "^2.2.5",
-    "@welldone-software/why-did-you-render": "^4.1.0-alpha.0",
     "axios": "^1.7.2",
     "chart.js": "^4.4.3",
     "color-namer": "^1.4.0",
-    "json-server": "^1.0.0-beta.0",
     "react": "^18.2.0",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.5",
     "react-loader-spinner": "^6.1.6",
