+++ random.txt
@@ -1,0 +1,18 @@
+
+ change avatar
+
+ return_requests--> fetch with return requests with totalPages,noOfRequests
+
+ orders:fetch orders with totalPages , noOfOrders
+
+ sellers:fetch sellers with  fullname,avatar,phoneNo,email
+
+ chart:days is x axis and revenue in y axis
+
+ fullname
+  avatar
+  products
+  role
+  address
+  phoneNo
+  email
