+++ src/components/Navbar.jsx
@@ -4,6 +4,7 @@ import { Bars3Icon, ShoppingCartIcon, XMarkIcon } from '@heroicons/react/24/outl
 import { Link, useNavigate } from 'react-router-dom'
 import { useFetchUserQuery, useLogoutUserMutation } from '../features/user/userSlice'
 import { catchAndShowMessage } from '../utils/catchAndShowMessage'
+import { Container } from './index'
 
 
 function classNames(...classes) {
@@ -85,85 +86,86 @@ export default function Navbar() {
                     </div>
                   </div>
                   {/* desktop cart and profile avatar */}
-                  {
-                    isAuthenticated && <div className="hidden md:block">
-                      <div className="ml-4 flex items-center md:ml-6">
-                        < Link to='/cart'
-                          className="relative rounded-full bg-gray-800 p-1 text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800 mr-2"
-                        >
-                          <span className="inline-flex items-center rounded-md bg-indigo-50 px-2 py-1 text-xs font-medium text-indigo-700 ring-1 ring-inset ring-indigo-700/10 absolute">266</span>
-                          <ShoppingCartIcon className="h-6 w-6 m-3" aria-hidden="true" />
-                        </Link>
+                  <Container
+                    RenderingConditions={[!!isAuthenticated]}
+                    className="hidden md:block"
+                  >
+                    <div className="ml-4 flex items-center md:ml-6">
+                      < Link to='/cart'
+                        className="relative rounded-full bg-gray-800 p-1 text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800 mr-2"
+                      >
+                        <span className="inline-flex items-center rounded-md bg-indigo-50 px-2 py-1 text-xs font-medium text-indigo-700 ring-1 ring-inset ring-indigo-700/10 absolute">266</span>
+                        <ShoppingCartIcon className="h-6 w-6 m-3" aria-hidden="true" />
+                      </Link>
 
-                        {/* Profile dropdown */}
-                        <Menu as="div" className="relative ml-3">
-                          <div>
-                            <MenuButton className="relative flex max-w-xs items-center rounded-full bg-gray-800 text-sm focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800">
-                              <span className="absolute -inset-1.5" />
-                              <span className="sr-only">Open user menu</span>
-                              <img className="h-8 w-8 rounded-full" src={userDetails.avatar} alt="" />
-                            </MenuButton>
-                          </div>
-                          <Transition
-                            as={Fragment}
-                            enter="transition ease-out duration-100"
-                            enterFrom="transform opacity-0 scale-95"
-                            enterTo="transform opacity-100 scale-100"
-                            leave="transition ease-in duration-75"
-                            leaveFrom="transform opacity-100 scale-100"
-                            leaveTo="transform opacity-0 scale-95"
-                          >
-                            <MenuItems className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
-                              {userNavigation.map((item) => (
-                                <MenuItem key={item.name}>
-                                  {({ active }) => (
-                                    <Link
-                                      to={item.to}
-                                      className={classNames(
-                                        active ? 'bg-gray-100' : '',
-                                        'block px-4 py-2 text-sm text-gray-700'
-                                      )}
-                                    >
-                                      {item.name}
-                                    </Link>
-                                  )}
-                                </MenuItem>
-                              ))}
-                              <MenuItem>
+                      {/* Profile dropdown */}
+                      <Menu as="div" className="relative ml-3">
+                        <div>
+                          <MenuButton className="relative flex max-w-xs items-center rounded-full bg-gray-800 text-sm focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800">
+                            <span className="absolute -inset-1.5" />
+                            <span className="sr-only">Open user menu</span>
+                            <img className="h-8 w-8 rounded-full" src={userDetails.avatar} alt="" />
+                          </MenuButton>
+                        </div>
+                        <Transition
+                          as={Fragment}
+                          enter="transition ease-out duration-100"
+                          enterFrom="transform opacity-0 scale-95"
+                          enterTo="transform opacity-100 scale-100"
+                          leave="transition ease-in duration-75"
+                          leaveFrom="transform opacity-100 scale-100"
+                          leaveTo="transform opacity-0 scale-95"
+                        >
+                          <MenuItems className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+                            {userNavigation.map((item) => (
+                              <MenuItem key={item.name}>
                                 {({ active }) => (
-                                  <button
-                                    type='button'
+                                  <Link
+                                    to={item.to}
                                     className={classNames(
                                       active ? 'bg-gray-100' : '',
-                                      'block px-4 py-2 text-sm text-gray-700 capitalize'
+                                      'block px-4 py-2 text-sm text-gray-700'
                                     )}
-                                    onClick={() => handleLogout()}
                                   >
-                                    sign out
-                                  </button>
+                                    {item.name}
+                                  </Link>
                                 )}
                               </MenuItem>
-                            </MenuItems>
-                          </Transition>
-                        </Menu>
-                      </div>
+                            ))}
+                            <MenuItem>
+                              {({ active }) => (
+                                <button
+                                  type='button'
+                                  className={classNames(
+                                    active ? 'bg-gray-100' : '',
+                                    'block px-4 py-2 text-sm text-gray-700 capitalize'
+                                  )}
+                                  onClick={() => catchAndShowMessage(LogoutUser)}
+                                >
+                                  sign out
+                                </button>
+                              )}
+                            </MenuItem>
+                          </MenuItems>
+                        </Transition>
+                      </Menu>
                     </div>
-                  }
+                  </Container>
                   {/* mobile hamburger and cross for open  */}
-                  {
-                    isAuthenticated && <div className="-mr-2 flex md:hidden">
-                      {/* Mobile menu button */}
-                      <DisclosureButton className="relative inline-flex items-center justify-center rounded-md bg-gray-800 p-2 text-gray-400 hover:bg-gray-700 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800">
-                        <span className="absolute -inset-0.5" />
-                        <span className="sr-only">Open main menu</span>
-                        {open ? (
-                          <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
-                        ) : (
-                          <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
-                        )}
-                      </DisclosureButton>
-                    </div>
-                  }
+                  <Container
+                    RenderingConditions={[!!isAuthenticated]}
+                    className="-mr-2 flex md:hidden"
+                  >
+                    <DisclosureButton className="relative inline-flex items-center justify-center rounded-md bg-gray-800 p-2 text-gray-400 hover:bg-gray-700 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800">
+                      <span className="absolute -inset-0.5" />
+                      <span className="sr-only">Open main menu</span>
+                      {open ? (
+                        <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
+                      ) : (
+                        <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
+                      )}
+                    </DisclosureButton>
+                  </Container>
                 </div>
               </div>
               {/* mobile navElements */}
@@ -185,46 +187,47 @@ export default function Navbar() {
                   ))}
 
                 </div>
-                {
-                  isAuthenticated && <div className="border-t border-gray-700 pb-3 pt-4">
-                    <div className="flex items-center px-5">
-                      <div className="flex-shrink-0">
-                        <img className="h-10 w-10 rounded-full" src={userDetails.avatar} alt="" />
-                      </div>
-                      <div className="ml-3">
-                        <div className="text-base font-medium leading-none text-white">{userDetails.fullname}</div>
-                        <div className="text-sm font-medium leading-none text-gray-400">{userDetails.email}</div>
-                      </div>
-                      <Link to='/cart'
-                        type="button"
-                        className="relative ml-auto flex-shrink-0 rounded-full bg-gray-800 p-1 text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800"
-                      >
-                        <span className="inline-flex items-center rounded-md bg-indigo-50 px-2 py-1 text-xs font-medium text-indigo-700 ring-1 ring-inset ring-indigo-700/10 absolute">266</span>
-                        <ShoppingCartIcon className="h-6 w-6 m-4" aria-hidden="true" />
-                      </Link>
+                <Container
+                  RenderingConditions={[!!isAuthenticated]}
+                  className="border-t border-gray-700 pb-3 pt-4"
+                >
+                  <div className="flex items-center px-5">
+                    <div className="flex-shrink-0">
+                      <img className="h-10 w-10 rounded-full" src={userDetails.avatar} alt="" />
                     </div>
-                    {/* profile setting and sign out link */}
-                    <div className="mt-3 space-y-1 px-2">
-                      {userNavigation.map((item) => (
-                        <DisclosureButton
-                          key={item.name}
-                          as="a"
-                          to={item.to}
-                          className="block rounded-md px-3 py-2 text-base font-medium text-gray-400 hover:bg-gray-700 hover:text-white"
-                        >
-                          {item.name}
-                        </DisclosureButton>
-                      ))}
-                      <DisclosureButton
-                        as="button"
-                        className="block w-full rounded-md px-3 py-2 text-base font-medium text-gray-400 hover:bg-gray-700 capitalize hover:text-white"
-                        onClick={() => catchAndShowMessage(LogoutUser)}
-                      >
-                        sign out
-                      </DisclosureButton>
+                    <div className="ml-3">
+                      <div className="text-base font-medium leading-none text-white">{userDetails.fullname}</div>
+                      <div className="text-sm font-medium leading-none text-gray-400">{userDetails.email}</div>
                     </div>
+                    <Link to='/cart'
+                      type="button"
+                      className="relative ml-auto flex-shrink-0 rounded-full bg-gray-800 p-1 text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800"
+                    >
+                      <span className="inline-flex items-center rounded-md bg-indigo-50 px-2 py-1 text-xs font-medium text-indigo-700 ring-1 ring-inset ring-indigo-700/10 absolute">266</span>
+                      <ShoppingCartIcon className="h-6 w-6 m-4" aria-hidden="true" />
+                    </Link>
                   </div>
-                }
+                  {/* profile setting and sign out link */}
+                  <div className="mt-3 space-y-1 px-2">
+                    {userNavigation.map((item) => (
+                      <DisclosureButton
+                        key={item.name}
+                        as="a"
+                        to={item.to}
+                        className="block rounded-md px-3 py-2 text-base font-medium text-gray-400 hover:bg-gray-700 hover:text-white"
+                      >
+                        {item.name}
+                      </DisclosureButton>
+                    ))}
+                    <DisclosureButton
+                      as="button"
+                      className="block w-full rounded-md px-3 py-2 text-base font-medium text-gray-400 hover:bg-gray-700 capitalize hover:text-white"
+                      onClick={() => catchAndShowMessage(LogoutUser)}
+                    >
+                      sign out
+                    </DisclosureButton>
+                  </div>
+                </Container>
               </DisclosurePanel>
             </>
           )}
