+++ src/features/cart/Cart.jsx
@@ -1,35 +1,52 @@
-import { Loader } from '../../components/index.js'
+import { useMemo } from 'react'
+import { Container } from '../../components/index.js'
 import { CartItem } from './components/CartItem'
 import { AmountSection } from './components/AmountSection'
-import { useCart } from '../../custom-hooks/useCart.js'
+import { useAddProductToCartMutation, useFetchUserCartQuery, useRemoveProductFromCartMutation } from './cartSlice.js'
 
 
 export default function Cart({ inCheckout = false }) {
-  // fetch cart
-  // update quantity
-  // remove item
 
-  const { userCart, isLoadingCart, subTotal, totalDiscount, AddToCart, RemoveFromCart } = useCart()
+
+  const [AddToCart, { isLoading: isLoadingAddToCart }] = useAddProductToCartMutation()
+
+  const [RemoveFromCart, { isLoading: isLoadingRemoveFromCart }] = useRemoveProductFromCartMutation()
+
+  const { data: { userCart = [] } = {}, isLoading: isLoadingCart,isFetching } = useFetchUserCartQuery()
+
+  console.log(isFetching);
+
+  const subTotal = useMemo(() => {
+    return userCart.reduce((acc, { price, quantity }) => acc + price * quantity, 0)
+  }, [userCart])
+
+  // total discount will be the memoized value in the useMemo fucntion
+  const totalDiscount = useMemo(() => {
+    return Math.round(userCart.reduce((acc, { price, quantity, discount }) => acc + price * discount * quantity / 100, 0));
+  }, [userCart])
 
   return (
-    cartStatus === 'loading' ? <Loader /> :
-      (userCart.length > 0 ? <div>
-        <div className="mt-8">
-          <h1 className='text-3xl font-bold mb-2'>{!inCheckout ? "Cart" : 'Your Items'}</h1>
-          <div className="flow-root">
-            <ul role="list" className="-my-6 divide-y divide-gray-200">
-              {
-                userCart.map((product) => (
-                  <CartItem product={product} AddToCart={AddToCart} RemoveFromCart={RemoveFromCart}
-                    key={product.product + product.size + product.color}
-                  />
-                ))
-              }
-            </ul>
-          </div>
+    <Container
+      LoadingConditions={[!!isLoadingCart, !!isLoadingAddToCart, !!isLoadingRemoveFromCart,!!isFetching]}
+      RenderingConditions={[!!userCart, !!userCart?.length > 0]}
+      backupElem={<h1 className='text-3xl capitalize text-center'>your cart is empty</h1>}
+    >
+      <div className="mt-8">
+        <h1 className='text-3xl font-bold mb-2'>{!inCheckout ? "Cart" : 'Your Items'}</h1>
+        <div className="flow-root">
+          <ul role="list" className="-my-6 divide-y divide-gray-200">
+            {
+              userCart.map((product) => (
+                <CartItem product={product} AddToCart={AddToCart} RemoveFromCart={RemoveFromCart}
+                  key={product.product + product.size + product.color}
+                />
+              ))
+            }
+          </ul>
         </div>
-        <AmountSection totalDiscount={totalDiscount} subTotal={subTotal} inCheckout={inCheckout} />
-      </div > : <h1 className='text-3xl capitalize text-center'>your cart is empty</h1>)
+      </div>
+      <AmountSection totalDiscount={totalDiscount} subTotal={subTotal} inCheckout={inCheckout} />
+    </Container>
   )
 }
