+++ src/features/dashboard/components/Orders.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Orders() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Orders
