+++ src/pages/ReturnDetailsPage.jsx
@@ -1,0 +1,12 @@
+import ReturnDetails from "../features/dashboard/components/ReturnRequestsDetails";
+import React from 'react'
+
+function ReturnDetailsPage() {
+    return (
+        <div>
+            <ReturnDetails />
+        </div>
+    )
+}
+
+export default ReturnDetailsPage
