+++ src/pages/ReturnRequestsPage.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import ReturnRequests from '../features/dashboard/components/ReturnRequests'
+
+const ReturnRequestsPage = () => {
+    return (
+        <div>
+            <ReturnRequests />
+        </div>
+    )
+}
+
+export default ReturnRequestsPage
