+++ src/pages/SellersPage.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Sellers from '../features/dashboard/components/Sellers'
+
+function SellersPage() {
+    return (
+        <div>
+            <Sellers />
+        </div>
+    )
+}
+
+export default SellersPage
